+++ CHANGELOG.md
@@ -4,17 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2020-11-2
+## [Unreleased] - 2020-11-3
 
 ### Added
 
 ### Changed
+* Refactored ParallelVariableAnnealingLength to eliminate redundant code
+* Refactored constructors of window limited permutation mutation operators to eliminate redundant code
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+* Fixed bug in window limited permutation mutation operators in handling of limit greater than or equal to permutation length.
+* Fixed missing parameter bounds check in LubyRestarts constructor.
 
 ### CI/CD